category create page

+++ pages/admin/productPage.js
@@ -31,14 +31,18 @@ const ProductPage = () => {
 
     const handleButtonDelete = async (e, id) => {
         e.preventDefault();
-        try {
-            await fetch("/api/category/deletedata?id=" + id, {
-                method: "DELETE",
-                headers: { "Content-Type": "application/json" },
-            })
-            window.location.reload()
-        } catch (err) {
-            console.log(err)
+        if(confirm('Are you sure want to delete this ?')){
+            try {
+                await fetch("/api/product/deletedata?id=" + id, {
+                    method: "DELETE",
+                    headers: { "Content-Type": "application/json" },
+                })
+                window.location.reload()
+            } catch (err) {
+                console.log(err)
+            }
+        }else{
+            console.log("not gonna happen!")
         }
     };
 
@@ -89,7 +93,7 @@ const ProductPage = () => {
             (
                 <>
                     <button>
-                        <Link href={`/admin/category/edit/${row.id}`}>
+                        <Link href={`/admin/product/edit/${row.id}`}>
                             <BiEdit className='h-5 w-5' />
                         </Link>
                     </button>
